fix: split registration helpers to satisfy react-refresh lint rule
Separate pure functions (validateSlug, passwordStrength) into
registration-utils.ts and keep only React components in
registration-helpers.tsx to satisfy the react-refresh/only-export-components
ESLint rule.
